refactor: useLiveComments を client/hooks/ に切り出す
live-comment.ts はデータ型 (LiveComment) とフック (useLiveComments) が混在して
いた。フックを client/hooks/use-live-comments.ts に分離し、型は lib に残す。型は
fixtures や表示コンポーネントが import するため lib が適切。
