+++ client/islands/Watch.tsx
@@ -1,7 +1,7 @@
 import { useEffect, useState } from "react";
 import type { components } from "../lib/api/schema.d.ts";
 import { $api } from "../lib/api/client.ts";
-import { useLiveComments } from "../lib/live-comment.ts";
+import { useLiveComments } from "../hooks/use-live-comments.ts";
 import { useNow } from "../hooks/use-now.ts";
 import type { ChannelType } from "../lib/service.ts";
 import { t } from "../locales/i18n.ts";
